Refactor use long double for spin calculations to improve precision

+++ include/spin/spin_calc.h
@@ -1,14 +1,14 @@
 #ifndef SPIN_CALC_H
 #define SPIN_CALC_H
 
-double spin_calc_r_dot_dot(double r, double theta, double theta_dot, double Phi_dot_0, double epsilon, double mass, double charge);
+long double spin_calc_r_dot_dot(long double r, long double theta, long double theta_dot, long double Phi_dot_0, long double epsilon, long double mass, long double charge);
 
-double spin_calc_theta_dot_dot(double r, double r_dot, double theta, double theta_dot, double phi_dot_0, double epsilon);
+long double spin_calc_theta_dot_dot(long double r, long double r_dot, long double theta, long double theta_dot, long double phi_dot_0, long double epsilon);
 
-double spin_calc_phi_dot(double phi_dot_0, double epsilon);
+long double spin_calc_phi_dot(long double phi_dot_0, long double epsilon);
 
-double spin_calc_phi_dot_0(double r, double mass, double nphi, double hbar, double theta);
+long double spin_calc_phi_dot_0(long double r, long double mass, long double nphi, long double hbar, long double theta);
 
-double spin_calc_epsilon(double r, double mass, double charge, double theta, double nphi);
+long double spin_calc_epsilon(long double r, long double mass, long double charge, long double theta, long double nphi);
 
 #endif // SPIN_CALC_H
